requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
flask-login==0.6.2
flask-wtf==1.1.1
flask-mail==0.9.1
pymysql==1.0.3
email-validator==2.0.0
python-dotenv==1.0.0
werkzeug==2.2.3
pyjwt==2.7.0
markdown==3.4.3
beautifulsoup4==4.12.2